+++ components/Home/sidebar/RootLeftSidebar.tsx
@@ -46,7 +46,8 @@ import InviteButton from "~/components/Referral/InviteButton";
 import killswitch from "~/config/killswitch/killswitch";
 import gateKeepCurrentUser from "~/config/gatekeeper/gateKeepCurrentUser";
 import RhTextTag from "~/components/shared/RhTextTag";
-import VerificationModal, { VerifiedBadge } from "~/components/Verification/VerificationModal";
+import VerificationModal from "~/components/Verification/VerificationModal";
+import VerifiedBadge from "~/components/Verification/VerifiedBadge";
 
 type Props = {
   openLoginModal: any;
@@ -373,16 +374,24 @@ function RootLeftSidebar({
         </div>
         <div className={css(styles.leftSidebarFooter)}>
           <div className={css(styles.leftSidebarFooterItemsTop)}>
-          <span className={css(formattedFooterTxtItem)}>
+            <span className={css(formattedFooterTxtItem)}>
               <VerificationModal
                 isModalOpen={isVerificationModalOpen}
                 handleModalClose={() => setIsVerificationModalOpen(false)}
               />
-              <span className={css(styles.referralProgramItem)} onClick={() => setIsVerificationModalOpen(true)} >
+              <span
+                className={css(styles.referralProgramItem)}
+                onClick={() => setIsVerificationModalOpen(true)}
+              >
                 {"Verify Authorship"}
-                <VerifiedBadge height={22} width={22} variation="grey" showTooltipOnHover={false} />
+                <VerifiedBadge
+                  height={22}
+                  width={22}
+                  variation="grey"
+                  showTooltipOnHover={false}
+                />
               </span>
-            </span>            
+            </span>
             <span className={css(formattedFooterTxtItem)}>
               <InviteButton context={"referral"}>
                 <span className={css(styles.referralProgramItem)}>
