+++ components/LiveFeed/Contribution/ContributionHeader.tsx
@@ -21,7 +21,7 @@ import ContentBadge from "~/components/ContentBadge";
 import { formatBountyAmount } from "~/config/types/bounty";
 import { truncateText } from "~/config/utils/string";
 import { COMMENT_TYPES } from "~/components/Comment/lib/types";
-import { VerifiedBadge } from "~/components/Verification/VerificationModal";
+import VerifiedBadge from "~/components/Verification/VerifiedBadge";
 
 type Args = {
   entry: Contribution;
@@ -175,11 +175,13 @@ const ContributionHeader = ({ entry }: Args) => {
       </div>
       <div className={css(styles.metadataWrapper)}>
         <div className={css(styles.metadataRow)}>
-          <div style={{
-            display: "flex",
-            alignItems: "center",
-            columnGap: "3px",
-          }}>
+          <div
+            style={{
+              display: "flex",
+              alignItems: "center",
+              columnGap: "3px",
+            }}
+          >
             <UserTooltip
               createdBy={createdBy}
               overrideTargetStyle={styles.userTooltip}
@@ -188,14 +190,14 @@ const ContributionHeader = ({ entry }: Args) => {
                   href={`/user/${createdBy.authorProfile?.id}/overview`}
                   key={`/user/${createdBy.authorProfile?.id}/overview-key`}
                 >
-                    {createdBy.authorProfile.firstName}{" "}
-                    {createdBy.authorProfile.lastName}
+                  {createdBy.authorProfile.firstName}{" "}
+                  {createdBy.authorProfile.lastName}
                 </ALink>
               }
             />
             {createdBy.authorProfile.isVerified && (
               <VerifiedBadge height={18} width={18} />
-            )}            
+            )}
 
             {actionLabel}
             {/* @ts-ignore */}
